+++ packages/fehu/examples/03-reinforce/main.ml
@@ -170,7 +170,7 @@ let () =
 
     let loss, grads = Rune.value_and_grad policy_tree loss_fn !params in
     let new_params, new_opt_state =
-      Vega.adam_step policy_tree ~lr !opt_state ~params:!params ~grads
+      Vega.adam_step policy_tree ~lr:(Vega.lr lr) !opt_state ~params:!params ~grads
     in
     params := new_params;
     opt_state := new_opt_state;